Funciones/Load_Business/requirements.txt
google-cloud-storage 
google-cloud-bigquery 
googlemaps 
pandas 
fsspec 
gcsfs 
pyarrow 
fastparquet 